+++ src/ila/event/minecraft/world/actor/item/ItemActorEvent.h
@@ -13,6 +13,8 @@ public:
     {
     }
 
+    void serialize(CompoundTag&) const override;
+
     ILNDAPI ItemActor& self() const;
 };