+++ src/ila/event/minecraft/world/level/block/actor/BlockActorEvent.h
@@ -1,0 +1,24 @@
+#include "ila/base/Macro.h"
+#include <ll/api/event/Event.h>
+#include <mc/world/level/block/actor/BlockActor.h>
+
+namespace ila::event::inline block::actor
+{
+
+class BlockActorEvent : public ll::event::Event
+{
+protected:
+    BlockActor& mSelf;
+
+public:
+    constexpr explicit BlockActorEvent(BlockActor& blockActor)
+        : mSelf(blockActor)
+    {
+    }
+
+    ILAPI void serialize(CompoundTag&) const override;
+
+    ILNDAPI BlockActor& self() const;
+};
+
+} // namespace ila::event::inline block::actor